+++ plugged/render-markdown.nvim/lua/render-markdown/core/treesitter.lua
@@ -1,5 +1,20 @@
+---@type table<string, vim.treesitter.Query>
+local queries = {}
+
 ---@class render.md.TreeSitter
 local M = {}
+
+---@param language string
+---@param query string
+---@return vim.treesitter.Query
+function M.parse(language, query)
+    local result = queries[query]
+    if result == nil then
+        result = vim.treesitter.query.parse(language, query)
+        queries[query] = result
+    end
+    return result
+end
 
 ---@param language string
 ---@param injection render.md.Injection?